frontend: add required paths to LLB and use it with --parents
This adds an additional `RequiredPaths` that is primarily intended for
use with `COPY --parents`. This parameter specifies expected directories
or files that should exist when performing the checksum. A not found
error will be produced if one of these paths is missing.

This fixes an issue with `COPY --parents` where a non existent directory
that was intended to be copied would be ignored.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>