solver: add ResolverCache support
New ResolverCache interface in JobContext allows build jobs to
memorize and synchronize accesses to mutable remote resources.

This is to make sure that when multiple parts of the same build
job, or build job and source metadata resolver access the same
remote resources, it remains the same for the duration of the
single build request, even if data happens to change on the remote side.

Fix such a possible case in the HTTP source. Even if the server
now returns completely different data, if the same URL was accessed
once for the ongoing build, then the initial contents are always
used until the build completes.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>