dockerfile: add subrequest for dumping LLB contents for dockerfile
This subrequest is similar to `buildctl debug dump-llb` but works on a
running buildkit daemon with the dockerfile frontend. The
`--call=dumpllb` option can be used to dump the raw LLB along with the
metadata and anything else in the definition to a JSON format.

This is useful for debugging purposes.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>