oci: Isolate cgroups under the current hierarchy
When buildkitd is run in a privileged container on Kubernetes, the
`/sys/fs/cgroup` mount will be that of the host which allows buildkitd
to remove itself from the cgroup hierarchy managed by Kubernetes
(cgroupfs). When buildkitd's worker creates cgroups outside of the
externally managed hierarchy, resource accounting is incorrect and
resource limits are not enforced. This can lead to OOM and other CPU
contention issues on nodes.

Introduce a new `isolateCgroups` configuration for the OCI worker. If
set, all cgroups are created beneath the cgroup hierarchy of the
buildkitd process.

Signed-off-by: Dan Duvall <dduvall@wikimedia.org>