linter: CopyIgnoredFile correctly cleans the path before attempting a match
When walking the filesystem, the walk would not have any leading paths
because it used `Walk("")` which didn't contain patterns like
`./myfile.txt`. When attempting to match a file path with the linter
rule, the initial section wouldn't be removed which would cause
`myfile.txt` not to match `./myfile.txt`.

For `CopyIgnoredFile`, this mostly affected things with exclusions since
a file path could match something very broad, like `**`, and then the
exclusion wouldn't match because it was too specific.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>