Fix markdown/links for AWS auth
The SDK link didn't end in a parenthesis, resulting in the markdown not
getting rendered as intended.

The resulting 'bare' go.dev link looked out of place, so moved it to a
more natural location.

Lastly, SDK link was for v1 docs, though the code actually uses v2; it
has been replaced with the v2 equivalent.

Signed-off-by: Roberto Villarreal <rrjjvv@yahoo.com>